requirements.txt
fastapi
uvicorn
python-multipart
openai
srt
requests
python-dotenv
pytest
pytest-asyncio
httpx
pytest-mock
pytest-cov
